labs/lab06-langgraph-stateful/requirements.txt
langchain>=0.1.0
langgraph>=0.0.30
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0